+++ test/doctests/test_titles.t
@@ -191,7 +191,12 @@ my %tests = (
 
 sub proc {
     my $file = shift;
-    convert2xml($file);
+    eval {
+        convert2xml($file);
+    };
+    if ($@) {
+        return ('', '');
+    }
     my $extractor = Extractor->new("$file.xml");
     $extractor->extract(qw/authors title/);
     system("rm $file.xml");
@@ -199,6 +204,7 @@ sub proc {
 }
 
 while (my ($file, $res) = each(%tests)) {
+    print substr($file, length('/home/wo/programming/opp-tools/test/doctests/')), "\n";
     my ($au, $ti) = proc($file);
     is(join(", ", @$au), join(", ", @{$res->{authors}}));
     is($ti, $res->{title});
